% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{fit_multi}
\alias{fit_multi}
\title{Fit multiple SVM's}
\usage{
fit_multi(data, tt_inds, tuning, k_cross = 10)
}
\arguments{
\item{data}{a data frame containing the data}

\item{tt_inds}{a vector of lists containing test and train indices (from tt_ranges())}

\item{tuning}{a tuning object (from tune_svm())}

\item{k_cross}{number of folds for cross validation, default is 10}
}
\value{
a list of objects of class svm from fit_svm each trained on a portion of the training data
}
\description{
Given data and a list of training and testing indices (from tt_ranges()) and a tuning object (from tune_svm()) will fit a SVM model to each testing set and return the resulting list of outputs
}
