% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_formulas}
\alias{get_formulas}
\title{Creates formulas from list}
\usage{
get_formulas(Cols, ind)
}
\arguments{
\item{Cols:}{vector of column names}

\item{ind:}{list of vectors, where each vector lists combination of columns to create formula from}
}
\value{
vector of strings ready to be passed to as.formula()
}
\description{
Given column names and combinations of columns wanted in model produces vector of strings that can be turned into formulae using as.formula()
}
