% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stonks.R
\name{import_stonks}
\alias{import_stonks}
\title{import_stonks}
\usage{
import_stonks(
  stock_outcome = c("BTC-USD"),
  stock_pred = c("ETH-USD"),
  lag = 1,
  from = "2018-01-01"
)
}
\arguments{
\item{stocks}{}
}
\description{
import_stonks
}
