% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{long_format}
\alias{long_format}
\title{Modify to long format}
\usage{
long_format(df)
}
\arguments{
\item{df, }{a data frame}
}
\value{
a data frame with added date column (if it was missing)
}
\description{
Checks if there is a column with dates, if not adds a column with dates from the row names (this is how data frames are created using import_stonks())
}
