% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotp}
\alias{plotp}
\title{Plot price action}
\usage{
plotp(data, ticker)
}
\arguments{
\item{data}{data frame containg price action data you want to plot}

\item{ticker}{string of the ticker you want to plot (must be one of the column names)}
}
\value{
Returns a plot of the price of the ticker against time
}
\description{
Plot price action
}
