% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{pred_multi}
\alias{pred_multi}
\title{Predict using multiple SVM's}
\usage{
pred_multi(data, tt_inds, svms)
}
\arguments{
\item{data}{a data frame containing the data}

\item{tt_inds}{a vector of lists containing test and train indices (from tt_ranges())}

\item{svms}{a list of svm objects (from fit_multi or just a list of results from fit_svm)}
}
\value{
a list of objects of class svm from fit_svm each trained on a portion of the training data
}
\description{
Given data and a list of training and testing indices (from tt_ranges()) and an svm object (from fit_svm()) will return predictions for testing data using their respective svm object
}
