% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{predict_svm}
\alias{predict_svm}
\title{Prediction}
\usage{
predict_svm(model, predictors, long = TRUE)
}
\arguments{
\item{model}{result from fit_svm, ie. object of class "svm"}

\item{predictors}{data frame containing data points for which we want predictions}

\item{long}{Boolean indicating if we want to pass the data frame to long_format before returning}
}
\value{
Data frame containing predictions
}
\description{
A wrapper for the prediction() function from the e1071 package. Calculates prediction then returns it in our desired format (a data.frame)
}
