% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMForecast-package.R
\docType{package}
\name{SVMForecast-package}
\alias{SVMForecast}
\alias{SVMForecast-package}
\title{SVMForecast: Forecasting yahoo finance data using SVM's}
\description{
Contains functions for getting price data from yahoo finance. Also contains functions for creating forecasting models for price data using SVM for regression.
}
\keyword{internal}
