% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotf}
\alias{plotf}
\title{Plot forecasting}
\usage{
plotf(data, train_range, test_range, preds)
}
\arguments{
\item{data}{data frame containing price action data you want to plot}

\item{train_range}{row indices used in training of SVM}

\item{test_range}{row indices used for testing of SVM}
}
\value{
Returns a plot
}
\description{
Will plot price action of BTC_USD for testing data (as a line), along with testing data (as black points) and the predictions (red dots).
}
